// client/src/typed-contracts/ManagerContract.d.ts
/* Generated by ts-generator ver. 0.0.8 */
/* tslint:disable */

import Contract, { CustomOptions, contractOptions } from "web3/eth/contract";
import { TransactionObject, BlockType } from "web3/eth/types";
import { Callback, EventLog } from "web3/types";
import { EventEmitter } from "events";
import { Provider } from "web3/providers";

export class ManagerContract {
  constructor(jsonInterface: any[], address?: string, options?: CustomOptions);
  _address: string;
  options: contractOptions;
  methods: {
    doesCategoryExist(arg0: string): TransactionObject<boolean>;

    categoryAddress(arg0: string | number[]): TransactionObject<string>;

    categoryContractsList(arg0: number | string): TransactionObject<string>;

    createVotingWithExistingCategory(
      _category: string,
      _question: string,
      _options: (string | number[])[],
      _votingEndTime: number | string,
      _resultsEndTime: number | string,
      _isPrivate: boolean,
      _permissions: (string)[]
    ): TransactionObject<void>;

    createVotingWithNewCategory(
      _categoryName: string | number[],
      _question: string,
      _options: (string | number[])[],
      _votingEndTime: number | string,
      _resultsEndTime: number | string,
      _isPrivate: boolean,
      _permissions: (string)[]
    ): TransactionObject<void>;

    numberOfCategories(): TransactionObject<string>;
  };
  deploy(options: {
    data: string;
    arguments: any[];
  }): TransactionObject<Contract>;
  events: {
    CategoryCreated(
      options?: {
        filter?: object;
        fromBlock?: BlockType;
        topics?: (null | string)[];
      },
      cb?: Callback<EventLog>
    ): EventEmitter;

    VotingCreated(
      options?: {
        filter?: object;
        fromBlock?: BlockType;
        topics?: (null | string)[];
      },
      cb?: Callback<EventLog>
    ): EventEmitter;

    allEvents: (
      options?: {
        filter?: object;
        fromBlock?: BlockType;
        topics?: (null | string)[];
      },
      cb?: Callback<EventLog>
    ) => EventEmitter;
  };
  getPastEvents(
    event: string,
    options?: {
      filter?: object;
      fromBlock?: BlockType;
      toBlock?: BlockType;
      topics?: (null | string)[];
    },
    cb?: Callback<EventLog[]>
  ): Promise<EventLog[]>;
  setProvider(provider: Provider): void;
}