// client/src/typed-contracts/VotingContract.d.ts
/* Generated by ts-generator ver. 0.0.8 */
/* tslint:disable */

import Contract, { CustomOptions, contractOptions } from "web3/eth/contract";
import { TransactionObject, BlockType } from "web3/eth/types";
import { Callback, EventLog } from "web3/types";
import { EventEmitter } from "events";
import { Provider } from "web3/providers";

export class VotingContract {
  constructor(jsonInterface: any[], address?: string, options?: CustomOptions);
  _address: string;
  options: contractOptions;
  methods: {
    hasVoted(arg0: string): TransactionObject<boolean>;

    options(arg0: number | string): TransactionObject<string>;

    viewContractInfo(): TransactionObject<{
      0: string;
      1: string;
      2: (string)[];
      3: string;
      4: string;
    }>;

    isPrivileged(user: string): TransactionObject<boolean>;

    vote(_option: number | string): TransactionObject<void>;

    votingEndTime(): TransactionObject<string>;
    question(): TransactionObject<string>;
    resultsEndTime(): TransactionObject<string>;
    category(): TransactionObject<string>;
    isPrivate(): TransactionObject<boolean>;
    numberOfOptions(): TransactionObject<string>;
    viewVotes(): TransactionObject<(string)[]>;
  };
  deploy(options: {
    data: string;
    arguments: any[];
  }): TransactionObject<Contract>;
  events: {
    allEvents: (
      options?: {
        filter?: object;
        fromBlock?: BlockType;
        topics?: (null | string)[];
      },
      cb?: Callback<EventLog>
    ) => EventEmitter;
  };
  getPastEvents(
    event: string,
    options?: {
      filter?: object;
      fromBlock?: BlockType;
      toBlock?: BlockType;
      topics?: (null | string)[];
    },
    cb?: Callback<EventLog[]>
  ): Promise<EventLog[]>;
  setProvider(provider: Provider): void;
}